Make screen cell's colour non-tiles-only
In one of those happy coincidences so rarely encountered in crawlcode,
it turns out that the cell->colour property is never used for rendering
tiles, only for console output. Conditionally adding that property
requires adding a lot of conditional compilation to view.cc's methods,
which has the nice side effect of revealing that these functions would
probably be a lot cleaner if they were entirely split in two.